Adding optional context to enumerate_parameters and friends (#39)
The `xcom_pull` on `BaseOperator` requires execution `context`, where that is baked into `TaskInstance`. This change allows the user to pass that context and then will pass that into `xcom_pull`. Because Python gets mad if you pass extra parameters, we are creating a param dictionary, adding the `context` when needed, and expanding that into the function.