Checking container status
Previously, we were only looking at the status of the pod. However, since the pod won't be marked as succeeded or failed until all the containers stop, the cloud_sql_proxy is preventing the 'failed' status from appearing. Instead, we will look at the individual containers and fail the operation if any of them are in a 'terminated' state with an exit code greater than zero.