Dev -> Master (#142)
# Adding tag support to stats

https://bluecore.atlassian.net/browse/PROD-13444

StatsD doesn't support tags, but we really want success/failure stats on a per task or per dag basis. This was accomplished by adding a tag-ignoring wrapper on top the StatsClient. Ironically, we now have the distinction of using a stats interface that matches neither statsd nor dogstatsd.

___NOTE:___ _WE DO NOT SUPPORT STATS IN DEV OR QA, SO THIS CANNOT BE TESTED!_

# Support for multiple connections in the KubernetesJobOperator sidecar

This isn't used in PROD yet, but is a backwards compatible chain